+++ Client/Frontend/Browser/TabCell.swift
@@ -4,6 +4,15 @@
 
 import Foundation
 import UIKit
+import Shared
+
+protocol TabTrayCell where Self: UICollectionViewCell {
+    /// True when the tab is the selected tab in the tray
+    var isSelectedTab: Bool { get }
+
+    /// Configure a tab cell using a Tab object, setting it's selected state at the same time
+    func configureWith(tab: Tab, isSelected selected: Bool)
+}
 
 class TabCell: UICollectionViewCell, TabTrayCell, ReusableCell {
 
@@ -246,6 +255,30 @@ class TabCell: UICollectionViewCell, TabTrayCell, ReusableCell {
             imageView.setImageAndBackground(forIcon: tab.displayFavicon, website: url) {}
             faviconBG.isHidden = false
             screenshotView.image = nil
+        }
+    }
+}
+
+extension TabTrayCell {
+
+    /// Use the display title unless it's an empty string, then use the base domain from the url
+    func getTabTrayTitle(tab: Tab) -> String? {
+        let baseDomain = tab.sessionData?.urls.last?.baseDomain ?? tab.url?.baseDomain
+        let urlLabel = baseDomain != nil ? baseDomain!.contains("local") ? "" : baseDomain : ""
+        return tab.displayTitle.isEmpty ? urlLabel : tab.displayTitle
+    }
+
+    func getA11yTitleLabel(tab: Tab) -> String? {
+        var aboutComponent = ""
+        if let url = tab.url, let about = InternalURL(url)?.aboutComponent { aboutComponent = about }
+        let baseName = tab.displayTitle.isEmpty ? aboutComponent.isEmpty ? getTabTrayTitle(tab: tab) : aboutComponent : tab.displayTitle
+
+        if isSelectedTab, let baseName = baseName, !baseName.isEmpty {
+            return baseName + ". " + String.TabTrayCurrentlySelectedTabAccessibilityLabel
+        } else if isSelectedTab {
+            return String.TabTrayCurrentlySelectedTabAccessibilityLabel
+        } else {
+            return baseName
         }
     }
 }
